requirements.txt
aiofiles==23.2.1
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
courlan==1.0.0
curl_cffi==0.6.2
cycler==0.12.1
dateparser==1.2.0
diskcache==5.6.3
duckduckgo_search==5.0
faiss-cpu==1.8.0
fastapi==0.109.2
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
googlesearch-python==1.2.3
gradio==4.19.2
gradio_client==0.10.1
h11==0.14.0
htmldate==1.7.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jusText==3.0.0
kiwisolver==1.4.5
langcodes==3.3.0
llama_cpp_python==0.2.26
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pefile==2023.2.7
pillow==10.2.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-multipart==0.0.9
pytweening==1.2.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
random-user-agent==1.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
ruff==0.2.2
safetensors==0.4.2
semantic-version==2.10.0
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.36.3
sympy==1.12
tld==0.13
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.0
torch==2.2.1
tqdm==4.66.1
trafilatura==1.7.0
transformers==4.38.2
txtai==7.0.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.25.0
websockets==11.0.3
